requirements.txt
# Optimized requirements for Dance Stick Figure (Python 3.10)
streamlit>=1.28.0
opencv-python>=4.8.0,<5.0.0
mediapipe>=0.10.0,<0.11.0
matplotlib>=3.7.0,<4.0.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0,<11.0.0

